requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
bcrypt==4.1.2
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
Mako==1.3.5
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
orjson==3.9.12
packaging==23.2
parso==0.8.3
passlib==1.7.4
platformdirs==4.1.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
Pygments==2.17.2
pytest==8.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.35.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
